mm: Stop kswapd early when nothing's waiting for it to free pages
Keeping kswapd running when all the failed allocations that invoked it
are satisfied incurs a high overhead due to unnecessary page eviction
and writeback, as well as spurious VM pressure events to various
registered shrinkers. When kswapd doesn't need to work to make an
allocation succeed anymore, stop it prematurely to save resources.

Change-Id: I0b64b5e90055c1bcf24628c1ebd576e0cbfa0b6d
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>